requirements.txt
requests
beautifulsoup4
python-dotenv
vpk

google-cloud-storage
